+++ lib/provider/gql/components/chat.dart
@@ -46,6 +46,8 @@ mixin ChatGraphQlMixin {
   ///
   /// Mandatory.
   Future<GetChat$Query> getChat(ChatId id) async {
+    Log.debug('getChat($id)', '$runtimeType');
+
     final variables = GetChatArguments(id: id);
     final QueryResult result = await client.query(
       QueryOptions(
@@ -94,6 +96,11 @@ mixin ChatGraphQlMixin {
     bool noFavorite = false,
     bool? withOngoingCalls,
   }) async {
+    Log.debug(
+      'recentChats($first, $after, $last, $before, $noFavorite, $withOngoingCalls)',
+      '$runtimeType',
+    );
+
     final variables = RecentChatsArguments(
       first: first,
       after: after,
@@ -144,6 +151,8 @@ mixin ChatGraphQlMixin {
     int? last,
     FavoriteChatsCursor? before,
   }) async {
+    Log.debug('favoriteChats($first, $after, $last, $before)', '$runtimeType');
+
     final variables = FavoriteChatsArguments(
       first: first,
       after: after,
@@ -174,6 +183,8 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op if a [Chat] with the given responder [User] exists
   /// already, and returns this [Chat].
   Future<ChatMixin> createDialogChat(UserId responderId) async {
+    Log.debug('createDialogChat($responderId)', '$runtimeType');
+
     final variables = CreateDialogArguments(responderId: responderId);
     final QueryResult result = await client.query(
       QueryOptions(
@@ -198,6 +209,8 @@ mixin ChatGraphQlMixin {
     List<UserId> memberIds, {
     ChatName? name,
   }) async {
+    Log.debug('createGroupChat($memberIds, $name)', '$runtimeType');
+
     final variables =
         CreateGroupChatArguments(memberIds: memberIds, name: name);
     final QueryResult result = await client.query(
@@ -236,7 +249,11 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified [Chat] has
   /// the specified name already
   Future<ChatEventsVersionedMixin?> renameChat(
-      ChatId id, ChatName? name) async {
+    ChatId id,
+    ChatName? name,
+  ) async {
+    Log.debug('renameChat($id, $name)', '$runtimeType');
+
     RenameChatArguments variables = RenameChatArguments(id: id, name: name);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -275,6 +292,11 @@ mixin ChatGraphQlMixin {
     ChatItemsCursor? before,
     bool onlyAttachments = false,
   }) async {
+    Log.debug(
+      'chatItems($id, $first, $after, $last, $before, $onlyAttachments)',
+      '$runtimeType',
+    );
+
     final variables = GetMessagesArguments(
       id: id,
       first: first,
@@ -324,6 +346,11 @@ mixin ChatGraphQlMixin {
     List<AttachmentId>? attachments,
     List<ChatItemId> repliesTo = const [],
   }) async {
+    Log.debug(
+      'postChatMessage($chatId, $text, $attachments, $repliesTo)',
+      '$runtimeType',
+    );
+
     final variables = PostChatMessageArguments(
       chatId: chatId,
       text: text,
@@ -363,7 +390,11 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified [User] is
   /// a member of the specified [Chat] already.
   Future<ChatEventsVersionedMixin?> addChatMember(
-      ChatId chatId, UserId userId) async {
+    ChatId chatId,
+    UserId userId,
+  ) async {
+    Log.debug('addChatMember($chatId, $userId)', '$runtimeType');
+
     final variables = AddChatMemberArguments(chatId: chatId, userId: userId);
     var result = await client.mutate(
       MutationOptions(
@@ -398,7 +429,11 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified [User] is
   /// not a member of the specified [Chat] already.
   Future<ChatEventsVersionedMixin?> removeChatMember(
-      ChatId chatId, UserId userId) async {
+    ChatId chatId,
+    UserId userId,
+  ) async {
+    Log.debug('removeChatMember($chatId, $userId)', '$runtimeType');
+
     RemoveChatMemberArguments variables =
         RemoveChatMemberArguments(chatId: chatId, userId: userId);
     var result = await client.mutate(
@@ -437,6 +472,8 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified [Chat] is
   /// already hidden by the authenticated [MyUser].
   Future<ChatEventsVersionedMixin?> hideChat(ChatId chatId) async {
+    Log.debug('hideChat($chatId)', '$runtimeType');
+
     HideChatArguments variables = HideChatArguments(chatId: chatId);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -483,6 +520,8 @@ mixin ChatGraphQlMixin {
     ChatId chatId,
     ChatItemId untilId,
   ) async {
+    Log.debug('readChat($chatId, $untilId)', '$runtimeType');
+
     final variables = ReadChatArguments(id: chatId, untilId: untilId);
     final QueryResult result = await client.query(
       QueryOptions(
@@ -542,6 +581,11 @@ mixin ChatGraphQlMixin {
     bool noFavorite = false,
     bool? withOngoingCalls,
   }) {
+    Log.debug(
+      'recentChatsTopEvents($count, $noFavorite, $withOngoingCalls)',
+      '$runtimeType',
+    );
+
     final variables = RecentChatsTopEventsArguments(
       count: count,
       noFavorite: noFavorite,
@@ -616,6 +660,8 @@ mixin ChatGraphQlMixin {
     ChatVersion? ver,
     FutureOr<ChatVersion?> Function() onVer,
   ) {
+    Log.debug('chatEvents($id, $ver, onVer)', '$runtimeType');
+
     final variables = ChatEventsArguments(id: id, ver: ver);
     return client.subscribe(
       SubscriptionOptions(
@@ -650,6 +696,8 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified [ChatItem]
   /// is hidden by the authenticated [MyUser] already.
   Future<ChatEventsVersionedMixin?> hideChatItem(ChatItemId id) async {
+    Log.debug('hideChatItem($id)', '$runtimeType');
+
     HideChatItemArguments variables = HideChatItemArguments(id: id);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -689,6 +737,8 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified
   /// [ChatMessage] is deleted already.
   Future<ChatEventsVersionedMixin?> deleteChatMessage(ChatItemId id) async {
+    Log.debug('deleteChatMessage($id)', '$runtimeType');
+
     DeleteChatMessageArguments variables = DeleteChatMessageArguments(id: id);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -729,6 +779,8 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified
   /// [ChatForward] is deleted already.
   Future<ChatEventsVersionedMixin?> deleteChatForward(ChatItemId id) async {
+    Log.debug('deleteChatForward($id)', '$runtimeType');
+
     DeleteChatForwardArguments variables = DeleteChatForwardArguments(id: id);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -765,6 +817,11 @@ mixin ChatGraphQlMixin {
     dio.MultipartFile? attachment, {
     void Function(int count, int total)? onSendProgress,
   }) async {
+    Log.debug(
+      'uploadAttachment($attachment, onSendProgress)',
+      '$runtimeType',
+    );
+
     final variables = UploadAttachmentArguments(file: null);
     final query = MutationOptions(
       operationName: 'UploadAttachment',
@@ -808,7 +865,7 @@ mixin ChatGraphQlMixin {
         );
       }
 
-      Log.print('[UploadAttachment] Failed with response: ${e.response}');
+      Log.error('Failed to upload attachment: ${e.response}', '$runtimeType');
 
       rethrow;
     }
@@ -836,6 +893,8 @@ mixin ChatGraphQlMixin {
   Future<ChatEventsVersionedMixin?> createChatDirectLink(
       ChatDirectLinkSlug slug,
       {ChatId? groupId}) async {
+    Log.debug('createChatDirectLink($slug, $groupId)', '$runtimeType');
+
     final variables =
         CreateChatDirectLinkArguments(slug: slug, groupId: groupId);
     final QueryResult result = await client.mutate(
@@ -868,8 +927,11 @@ mixin ChatGraphQlMixin {
   ///
   /// Succeeds as no-op (and returns no [ChatEvent]) if the given [Chat]-group
   /// has no active [ChatDirectLink]s already.
-  Future<ChatEventsVersionedMixin?> deleteChatDirectLink(
-      {ChatId? groupId}) async {
+  Future<ChatEventsVersionedMixin?> deleteChatDirectLink({
+    ChatId? groupId,
+  }) async {
+    Log.debug('deleteChatDirectLink($groupId)', '$runtimeType');
+
     final variables = DeleteChatDirectLinkArguments(groupId: groupId);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -907,6 +969,8 @@ mixin ChatGraphQlMixin {
   /// the [Chat]-dialog by the specified [ChatDirectLink].
   Future<UseChatDirectLink$Mutation$UseChatDirectLink$UseChatDirectLinkOk>
       useChatDirectLink(ChatDirectLinkSlug slug) async {
+    Log.debug('useChatDirectLink($slug)', '$runtimeType');
+
     final variables = UseChatDirectLinkArguments(slug: slug);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -956,6 +1020,8 @@ mixin ChatGraphQlMixin {
   /// - The server is shutting down or becoming unreachable (unexpectedly
   /// completes after initialization)
   Stream<QueryResult> keepTyping(ChatId id) {
+    Log.debug('keepTyping($id)', '$runtimeType');
+
     final variables = KeepTypingArguments(chatId: id);
     return client.subscribe(
       SubscriptionOptions(
@@ -993,6 +1059,11 @@ mixin ChatGraphQlMixin {
     ChatMessageAttachmentsInput? attachments,
     ChatMessageRepliesInput? repliesTo,
   }) async {
+    Log.debug(
+      'editChatMessage($id, $text, $attachments, $repliesTo)',
+      '$runtimeType',
+    );
+
     final variables = EditChatMessageArguments(
       id: id,
       text: text,
@@ -1046,6 +1117,11 @@ mixin ChatGraphQlMixin {
     ChatMessageText? text,
     List<AttachmentId>? attachments,
   }) async {
+    Log.debug(
+      'forwardChatItems($from, $to, $items, $text, $attachments)',
+      '$runtimeType',
+    );
+
     final variables = ForwardChatItemsArguments(
       from: from,
       to: to,
@@ -1100,6 +1176,8 @@ mixin ChatGraphQlMixin {
     ChatId id,
     Muting? mute,
   ) async {
+    Log.debug('toggleChatMute($id, $mute)', '$runtimeType');
+
     final variables = ToggleChatMuteArguments(id: id, mute: mute);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -1125,6 +1203,8 @@ mixin ChatGraphQlMixin {
   ///
   /// Mandatory.
   Future<GetAttachments$Query> attachments(ChatItemId id) async {
+    Log.debug('attachments($id)', '$runtimeType');
+
     final variables = GetAttachmentsArguments(id: id);
     final QueryResult result = await client.query(
       QueryOptions(
@@ -1162,6 +1242,11 @@ mixin ChatGraphQlMixin {
     dio.MultipartFile? file,
     void Function(int count, int total)? onSendProgress,
   }) async {
+    Log.debug(
+      'updateChatAvatar($id, $file, onSendProgress)',
+      '$runtimeType',
+    );
+
     final variables = UpdateChatAvatarArguments(chatId: id, file: null);
     final query = MutationOptions(
       operationName: 'UpdateChatAvatar',
@@ -1239,6 +1324,8 @@ mixin ChatGraphQlMixin {
     ChatId id,
     ChatFavoritePosition position,
   ) async {
+    Log.debug('favoriteChat($id, $position)', '$runtimeType');
+
     final variables = FavoriteChatArguments(id: id, pos: position);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -1272,6 +1359,8 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op (and returns no [ChatEvent]) if the specified [Chat] is
   /// not in the favorites list already.
   Future<ChatEventsVersionedMixin?> unfavoriteChat(ChatId id) async {
+    Log.debug('unfavoriteChat($id)', '$runtimeType');
+
     final variables = UnfavoriteChatArguments(id: id);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -1340,6 +1429,8 @@ mixin ChatGraphQlMixin {
   Stream<QueryResult> favoriteChatsEvents(
     FavoriteChatsListVersion? Function() ver,
   ) {
+    Log.debug('favoriteChatsEvents(ver)', '$runtimeType');
+
     final variables = FavoriteChatsEventsArguments(ver: ver());
     return client.subscribe(
       SubscriptionOptions(
@@ -1372,6 +1463,8 @@ mixin ChatGraphQlMixin {
     ChatId id,
     ChatItemId untilId,
   ) async {
+    Log.debug('clearChat($id, $untilId)', '$runtimeType');
+
     final ClearChatArguments variables =
         ClearChatArguments(id: id, untilId: untilId);
     final QueryResult result = await client.mutate(
@@ -1402,6 +1495,8 @@ mixin ChatGraphQlMixin {
   /// Succeeds as no-op if the [Chat]-monolog for the authenticated [MyUser]
   /// exists already, and returns it.
   Future<ChatMixin> createMonologChat(ChatName? name) async {
+    Log.debug('createMonologChat($name)', '$runtimeType');
+
     final variables = CreateMonologChatArguments(name: name);
     final QueryResult result = await client.mutate(
       MutationOptions(
@@ -1427,6 +1522,8 @@ mixin ChatGraphQlMixin {
   /// Query returns `null` when no [Chat]-monolog exists for the authenticated
   /// [MyUser].
   Future<ChatMixin?> getMonolog() async {
+    Log.debug('getMonolog()', '$runtimeType');
+
     final QueryResult result = await client.query(
       QueryOptions(
         operationName: 'GetMonolog',