+++ lib/store/my_user.dart
@@ -43,6 +43,7 @@ import '/provider/hive/blocklist.dart';
 import '/provider/hive/my_user.dart';
 import '/provider/hive/user.dart';
 import '/util/backoff.dart';
+import '/util/log.dart';
 import '/util/new_type.dart';
 import '/util/platform_utils.dart';
 import '/util/stream_utils.dart';
@@ -109,6 +110,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     required Function() onUserDeleted,
     required Function() onPasswordUpdated,
   }) async {
+    Log.debug('init(onUserDeleted, onPasswordUpdated)', '$runtimeType');
+
     this.onPasswordUpdated = onPasswordUpdated;
     this.onUserDeleted = onUserDeleted;
 
@@ -163,6 +166,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   void dispose() {
+    Log.debug('dispose()', '$runtimeType');
+
     _localSubscription?.cancel();
     _blocklistSubscription?.cancel();
     _remoteSubscription?.close(immediate: true);
@@ -172,10 +177,15 @@ class MyUserRepository implements AbstractMyUserRepository {
   }
 
   @override
-  Future<void> clearCache() => _myUserLocal.clear();
+  Future<void> clearCache() async {
+    Log.debug('clearCache()', '$runtimeType');
+    await _myUserLocal.clear();
+  }
 
   @override
   Future<void> updateUserName(UserName? name) async {
+    Log.debug('updateUserName($name)', '$runtimeType');
+
     final UserName? oldName = myUser.value?.name;
 
     myUser.update((u) => u?.name = name);
@@ -190,6 +200,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> updateUserStatus(UserTextStatus? status) async {
+    Log.debug('updateUserStatus($status)', '$runtimeType');
+
     final UserTextStatus? oldStatus = myUser.value?.status;
 
     myUser.update((u) => u?.status = status);
@@ -204,6 +216,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> updateUserLogin(UserLogin login) async {
+    Log.debug('updateUserLogin($login)', '$runtimeType');
+
     final UserLogin? oldLogin = myUser.value?.login;
 
     myUser.update((u) => u?.login = login);
@@ -218,6 +232,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> updateUserPresence(Presence presence) async {
+    Log.debug('updateUserPresence($presence)', '$runtimeType');
+
     final Presence? oldPresence = myUser.value?.presence;
 
     myUser.update((u) => u?.presence = presence);
@@ -235,6 +251,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     UserPassword? oldPassword,
     UserPassword newPassword,
   ) async {
+    Log.debug('updateUserPassword(***, ***)', '$runtimeType');
+
     final bool? hasPassword = myUser.value?.hasPassword;
 
     myUser.update((u) => u?.hasPassword = true);
@@ -251,10 +269,15 @@ class MyUserRepository implements AbstractMyUserRepository {
   }
 
   @override
-  Future<void> deleteMyUser() => _graphQlProvider.deleteMyUser();
+  Future<void> deleteMyUser() async {
+    Log.debug('deleteMyUser()', '$runtimeType');
+    await _graphQlProvider.deleteMyUser();
+  }
 
   @override
   Future<void> deleteUserEmail(UserEmail email) async {
+    Log.debug('deleteUserEmail($email)', '$runtimeType');
+
     if (myUser.value?.emails.unconfirmed == email) {
       final UserEmail? unconfirmed = myUser.value?.emails.unconfirmed;
 
@@ -287,6 +310,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> deleteUserPhone(UserPhone phone) async {
+    Log.debug('deleteUserPhone($phone)', '$runtimeType');
+
     if (myUser.value?.phones.unconfirmed == phone) {
       final UserPhone? unconfirmed = myUser.value?.phones.unconfirmed;
 
@@ -319,6 +344,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> addUserEmail(UserEmail email) async {
+    Log.debug('addUserEmail($email)', '$runtimeType');
+
     final UserEmail? unconfirmed = myUser.value?.emails.unconfirmed;
 
     myUser.update((u) => u?.emails.unconfirmed = email);
@@ -333,6 +360,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> addUserPhone(UserPhone phone) async {
+    Log.debug('addUserPhone($phone)', '$runtimeType');
+
     final UserPhone? unconfirmed = myUser.value?.phones.unconfirmed;
 
     myUser.update((u) => u?.phones.unconfirmed = phone);
@@ -347,6 +376,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> confirmEmailCode(ConfirmationCode code) async {
+    Log.debug('confirmEmailCode($code)', '$runtimeType');
+
     final UserEmail? unconfirmed = myUser.value?.emails.unconfirmed;
 
     await _graphQlProvider.confirmEmailCode(code);
@@ -364,6 +395,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> confirmPhoneCode(ConfirmationCode code) async {
+    Log.debug('confirmPhoneCode($code)', '$runtimeType');
+
     final UserPhone? unconfirmed = myUser.value?.phones.unconfirmed;
 
     await _graphQlProvider.confirmPhoneCode(code);
@@ -380,13 +413,21 @@ class MyUserRepository implements AbstractMyUserRepository {
   }
 
   @override
-  Future<void> resendEmail() => _graphQlProvider.resendEmail();
+  Future<void> resendEmail() async {
+    Log.debug('resendEmail()', '$runtimeType');
+    await _graphQlProvider.resendEmail();
+  }
 
   @override
-  Future<void> resendPhone() => _graphQlProvider.resendPhone();
+  Future<void> resendPhone() async {
+    Log.debug('resendPhone()', '$runtimeType');
+    await _graphQlProvider.resendPhone();
+  }
 
   @override
   Future<void> createChatDirectLink(ChatDirectLinkSlug slug) async {
+    Log.debug('createChatDirectLink($slug)', '$runtimeType');
+
     final ChatDirectLink? link = myUser.value?.chatDirectLink;
 
     myUser.update((u) => u?.chatDirectLink = ChatDirectLink(slug: slug));
@@ -401,6 +442,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> deleteChatDirectLink() async {
+    Log.debug('deleteChatDirectLink()', '$runtimeType');
+
     final ChatDirectLink? link = myUser.value?.chatDirectLink;
 
     myUser.update((u) => u?.chatDirectLink = null);
@@ -418,6 +461,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     NativeFile? file, {
     void Function(int count, int total)? onSendProgress,
   }) async {
+    Log.debug('updateAvatar($file, onSendProgress)', '$runtimeType');
+
     dio.MultipartFile? upload;
 
     if (file != null) {
@@ -470,6 +515,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> toggleMute(MuteDuration? mute) async {
+    Log.debug('toggleMute($mute)', '$runtimeType');
+
     final MuteDuration? muted = myUser.value?.muted;
 
     final Muting? muting = mute == null
@@ -491,6 +538,8 @@ class MyUserRepository implements AbstractMyUserRepository {
     NativeFile? file, {
     void Function(int count, int total)? onSendProgress,
   }) async {
+    Log.debug('updateCallCover($file, onSendProgress)', '$runtimeType');
+
     dio.MultipartFile? upload;
 
     if (file != null) {
@@ -543,6 +592,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   @override
   Future<void> refresh() async {
+    Log.debug('refresh()', '$runtimeType');
+
     final response = await _graphQlProvider.getMyUser();
 
     if (response.myUser != null) {
@@ -554,6 +605,8 @@ class MyUserRepository implements AbstractMyUserRepository {
   //       loading on demand.
   /// Fetches __all__ blacklisted [User]s from the remote.
   Future<List<HiveUser>> _fetchBlocklist() async {
+    Log.debug('_fetchBlocklist()', '$runtimeType');
+
     final query = await _graphQlProvider.getBlocklist(first: 120);
     final users = query.edges.map((e) => e.node.user.toHive()).toList();
     users.forEach(_userRepo.put);
@@ -563,6 +616,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Initializes [MyUserHiveProvider.boxEvents] subscription.
   Future<void> _initLocalSubscription() async {
+    Log.debug('_initLocalSubscription()', '$runtimeType');
+
     _localSubscription = StreamIterator(_myUserLocal.boxEvents);
     while (await _localSubscription!.moveNext()) {
       BoxEvent event = _localSubscription!.current;
@@ -581,6 +636,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Initializes [BlocklistHiveProvider.boxEvents] subscription.
   Future<void> _initBlacklistSubscription() async {
+    Log.debug('_initBlacklistSubscription()', '$runtimeType');
+
     _blocklistSubscription = StreamIterator(_blocklistLocal.boxEvents);
     while (await _blocklistSubscription!.moveNext()) {
       final BoxEvent event = _blocklistSubscription!.current;
@@ -601,6 +658,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Initializes [_myUserRemoteEvents] subscription.
   Future<void> _initRemoteSubscription() async {
+    Log.debug('_initRemoteSubscription()', '$runtimeType');
+
     _remoteSubscription?.close(immediate: true);
     _remoteSubscription =
         StreamQueue(_myUserRemoteEvents(() => _myUserLocal.myUser?.ver));
@@ -609,6 +668,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Initializes the [GraphQlProvider.keepOnline] subscription.
   void _initKeepOnlineSubscription() {
+    Log.debug('_initKeepOnlineSubscription()', '$runtimeType');
+
     _keepOnlineSubscription?.cancel();
     _keepOnlineSubscription = _graphQlProvider.keepOnline().listen(
       (_) {
@@ -622,6 +683,8 @@ class MyUserRepository implements AbstractMyUserRepository {
 
   /// Saves the provided [user] in [Hive].
   void _setMyUser(HiveMyUser user, {bool ignoreVersion = false}) {
+    Log.debug('_setMyUser($user, $ignoreVersion)', '$runtimeType');
+
     if (user.ver > _myUserLocal.myUser?.ver || ignoreVersion) {
       _myUserLocal.set(user);
     }
@@ -632,9 +695,18 @@ class MyUserRepository implements AbstractMyUserRepository {
     var userEntity = _myUserLocal.myUser;
 
     if (userEntity == null || versioned.ver <= userEntity.ver) {
+      Log.debug(
+        '_myUserRemoteEvent(): ignored ${versioned.events.map((e) => e.kind)}',
+        '$runtimeType',
+      );
       return;
     }
     userEntity.ver = versioned.ver;
+
+    Log.debug(
+      '_myUserRemoteEvent(): ${versioned.events.map((e) => e.kind)}',
+      '$runtimeType',
+    );
 
     for (var event in versioned.events) {
       // Updates a [User] associated with this [MyUserEvent.userId].
@@ -824,28 +896,44 @@ class MyUserRepository implements AbstractMyUserRepository {
   /// Subscribes to remote [MyUserEvent]s of the authenticated [MyUser].
   Stream<MyUserEventsVersioned> _myUserRemoteEvents(
     MyUserVersion? Function() ver,
-  ) =>
-      _graphQlProvider.myUserEvents(ver).asyncExpand((event) async* {
-        var events =
-            MyUserEvents$Subscription.fromJson(event.data!).myUserEvents;
+  ) {
+    Log.debug('_myUserRemoteEvents(ver)', '$runtimeType');
 
-        if (events.$$typename == 'SubscriptionInitialized') {
-          events
-              as MyUserEvents$Subscription$MyUserEvents$SubscriptionInitialized;
-          // No-op.
-        } else if (events.$$typename == 'MyUser') {
-          _setMyUser((events as MyUserMixin).toHive());
-        } else if (events.$$typename == 'MyUserEventsVersioned') {
-          var mixin = events as MyUserEventsVersionedMixin;
-          yield MyUserEventsVersioned(
-            mixin.events.map((e) => _myUserEvent(e)).toList(),
-            mixin.ver,
-          );
-        }
-      });
+    return _graphQlProvider.myUserEvents(ver).asyncExpand((event) async* {
+      Log.trace('_myUserRemoteEvents(ver): ${event.data}', '$runtimeType');
+
+      var events = MyUserEvents$Subscription.fromJson(event.data!).myUserEvents;
+
+      if (events.$$typename == 'SubscriptionInitialized') {
+        Log.debug(
+          '_myUserRemoteEvents(ver): SubscriptionInitialized',
+          '$runtimeType',
+        );
+
+        events
+            as MyUserEvents$Subscription$MyUserEvents$SubscriptionInitialized;
+        // No-op.
+      } else if (events.$$typename == 'MyUser') {
+        Log.debug(
+          '_myUserRemoteEvents(ver): MyUser',
+          '$runtimeType',
+        );
+
+        _setMyUser((events as MyUserMixin).toHive());
+      } else if (events.$$typename == 'MyUserEventsVersioned') {
+        var mixin = events as MyUserEventsVersionedMixin;
+        yield MyUserEventsVersioned(
+          mixin.events.map((e) => _myUserEvent(e)).toList(),
+          mixin.ver,
+        );
+      }
+    });
+  }
 
   /// Constructs a [MyUserEvent] from the [MyUserEventsVersionedMixin$Events].
   MyUserEvent _myUserEvent(MyUserEventsVersionedMixin$Events e) {
+    Log.trace('_myUserEvent($e)', '$runtimeType');
+
     if (e.$$typename == 'EventUserNameUpdated') {
       var node = e as MyUserEventsVersionedMixin$Events$EventUserNameUpdated;
       return EventUserNameUpdated(node.userId, node.name);