+++ lib/ui/page/call/component/common.dart
@@ -56,7 +56,7 @@ abstract class CallButton {
 
 /// [CallButton] toggling a more panel.
 class MoreButton extends CallButton {
-  const MoreButton(CallController c) : super(c);
+  const MoreButton(super.c);
 
   @override
   bool get isRemovable => false;
@@ -80,7 +80,7 @@ class MoreButton extends CallButton {
 
 /// [CallButton] toggling a local video.
 class VideoButton extends CallButton {
-  const VideoButton(CallController c) : super(c);
+  const VideoButton(super.c);
 
   @override
   String get hint {
@@ -122,7 +122,7 @@ class VideoButton extends CallButton {
 
 /// [CallButton] toggling a local audio.
 class AudioButton extends CallButton {
-  const AudioButton(CallController c) : super(c);
+  const AudioButton(super.c);
 
   @override
   String get hint {
@@ -164,7 +164,7 @@ class AudioButton extends CallButton {
 
 /// [CallButton] toggling a local screen-sharing.
 class ScreenButton extends CallButton {
-  const ScreenButton(CallController c) : super(c);
+  const ScreenButton(super.c);
 
   @override
   String get hint {
@@ -201,7 +201,7 @@ class ScreenButton extends CallButton {
 
 /// [CallButton] toggling hand.
 class HandButton extends CallButton {
-  const HandButton(CallController c) : super(c);
+  const HandButton(super.c);
 
   @override
   String get hint => c.isMobile
@@ -232,7 +232,7 @@ class HandButton extends CallButton {
 
 /// [CallButton] invoking the [CallController.openSettings].
 class SettingsButton extends CallButton {
-  const SettingsButton(CallController c) : super(c);
+  const SettingsButton(super.c);
 
   @override
   String get hint => 'btn_call_settings'.l10n;
@@ -253,7 +253,7 @@ class SettingsButton extends CallButton {
 
 /// [CallButton] invoking the [CallController.openAddMember].
 class ParticipantsButton extends CallButton {
-  const ParticipantsButton(CallController c) : super(c);
+  const ParticipantsButton(super.c);
 
   @override
   String get hint =>
@@ -276,7 +276,7 @@ class ParticipantsButton extends CallButton {
 
 /// [CallButton] toggling the remote video.
 class RemoteVideoButton extends CallButton {
-  const RemoteVideoButton(CallController c) : super(c);
+  const RemoteVideoButton(super.c);
 
   @override
   String get hint => c.isMobile
@@ -307,7 +307,7 @@ class RemoteVideoButton extends CallButton {
 
 /// [CallButton] toggling the remote audio.
 class RemoteAudioButton extends CallButton {
-  const RemoteAudioButton(CallController c) : super(c);
+  const RemoteAudioButton(super.c);
 
   @override
   String get hint => c.isMobile
@@ -401,7 +401,7 @@ class AcceptVideoButton extends CallButton {
 
 /// [RoundFloatingButton] declining a call.
 class DeclineButton extends CallButton {
-  const DeclineButton(CallController c) : super(c);
+  const DeclineButton(super.c);
 
   @override
   String get hint => 'btn_call_decline'.l10n;
@@ -426,7 +426,7 @@ class DeclineButton extends CallButton {
 
 /// [RoundFloatingButton] canceling an outgoing call.
 class CancelButton extends CallButton {
-  const CancelButton(CallController c) : super(c);
+  const CancelButton(super.c);
 
   @override
   String get hint => 'btn_call_cancel'.l10n;
@@ -456,7 +456,7 @@ class CancelButton extends CallButton {
 
 /// [CallButton] ending the call.
 class EndCallButton extends CallButton {
-  const EndCallButton(CallController c) : super(c);
+  const EndCallButton(super.c);
 
   @override
   bool get isRemovable => false;
@@ -484,7 +484,7 @@ class EndCallButton extends CallButton {
 
 /// [RoundFloatingButton] switching a speaker output.
 class SpeakerButton extends CallButton {
-  const SpeakerButton(CallController c) : super(c);
+  const SpeakerButton(super.c);
 
   @override
   String get hint => c.isMobile
@@ -517,7 +517,7 @@ class SpeakerButton extends CallButton {
 
 /// [RoundFloatingButton] switching a local video stream.
 class SwitchButton extends CallButton {
-  const SwitchButton(CallController c) : super(c);
+  const SwitchButton(super.c);
 
   @override
   String get hint => c.isMobile