+++ lib/ui/page/home/tab/chats/widget/recent_chat.dart
@@ -771,6 +771,7 @@ class RecentChatTile extends StatelessWidget {
         final bool isSent = item.status.value == SendingStatus.sent;
         final bool isRead =
             chat.members.length <= 1 ? isSent : chat.isRead(item, me) && isSent;
+        final bool isHalfRead = chat.isHalfRead(item, me);
         final bool isDelivered = isSent && !chat.lastDelivery.isBefore(item.at);
         final bool isError = item.status.value == SendingStatus.error;
         final bool isSending = item.status.value == SendingStatus.sending;
@@ -779,9 +780,13 @@ class RecentChatTile extends StatelessWidget {
           padding: const EdgeInsets.only(right: 4),
           child: SvgIcon(
             isRead
-                ? inverted
-                    ? SvgIcons.readWhite
-                    : SvgIcons.read
+                ? isHalfRead
+                    ? inverted
+                        ? SvgIcons.halfReadWhite
+                        : SvgIcons.halfRead
+                    : inverted
+                        ? SvgIcons.readWhite
+                        : SvgIcons.read
                 : isDelivered
                     ? inverted
                         ? SvgIcons.deliveredWhite