+++ CHANGELOG.md
@@ -18,6 +18,8 @@ All user visible changes to this project will be documented in this file. This p
     - Download and Save as option from context menu. ([#697], [#654])
     - Style page:
         - Redesigned widgets tab. ([#695], [#632])
+    - Chat page:
+        - Display read partially message status in groups. ([#703], [#666])
 
 ### Fixed
 
@@ -25,10 +27,12 @@ All user visible changes to this project will be documented in this file. This p
     - Invalid caller name in popup calls. ([#711])
 
 [#632]: /../../issues/632
+[#666]: /../../issues/666
 [#681]: /../../issues/681
 [#654]: /../../issues/654
 [#695]: /../../pull/695
 [#702]: /../../pull/702
+[#703]: /../../pull/703
 [#711]: /../../pull/711
 [#697]: /../../pull/697